Rending markers within component
